✨ feat(footer): add copyright information and navigation menu
The code author added a new HTML file called footer.component.html to the src/app/footer directory. This file contains the code for the footer component, which includes a copyright section and a navigation menu.

The footer component displays the copyright year (2023) and the author's name (Evan Viguié). It also includes a link to the terms of use and a button to return to the homepage.

This change improves the user experience by providing relevant information and easy navigation options in the footer.

Issue: #1234 (master)