✨ Upgrade @angular/cdk and @angular/material to version 17.0.1
The code author wanted to upgrade the versions of @angular/cdk and @angular/material libraries to version 17.0.1. This upgrade will bring in bug fixes, improvements, and possibly new features to the codebase. The previous versions were @angular/cdk 17.0.0 and @angular/material 17.0.0.

This upgrade will ensure that the codebase is up-to-date with the latest versions of these libraries and take advantage of any new features or improvements that have been introduced. It will also address any known issues or bugs that have been fixed in the new versions.

This commit will help in maintaining a stable and reliable codebase by keeping the dependencies updated. It will also ensure that the codebase is compatible with other libraries and frameworks that may rely on these versions of @angular/cdk and @angular/material.

Issue: N/A (master)