🔀 Add leaflet dependency to package.json (master)
This commit adds the leaflet dependency to the package.json file. Leaflet is a JavaScript library for interactive maps. It allows us to integrate maps into our Angular application. By adding this dependency, we can leverage the features and functionality provided by Leaflet to enhance the user experience of our application.

The package-lock.json file is also updated with the new leaflet version and its corresponding integrity.

Changes:
- package.json: Added "leaflet": "^1.9.4" to the dependencies section.
- package-lock.json: Added leaflet's version and integrity information.

This change is made to improve the mapping capabilities of our application and provide a more interactive and engaging experience for our users.

Issue: N/A