✨:sparkles: Add USA capitals geojson data (master)
This commit adds a new file `usa-capitals.geojson` to the `src/assets/data`
directory. The file contains detailed information about the capitals of each
state in the United States, including their coordinates and population.

The addition of this file will be beneficial for various features that require
access to this information, such as displaying the location of each capital on a
map or providing population statistics.