✨ (feat): Add map component with Leaflet integration (master)
This commit adds a new file `map.component.ts` which contains the implementation of the `MapComponent`. The component is responsible for displaying a map using the Leaflet library in an Angular application.

The `MapComponent` class imports the necessary dependencies, including `L` from Leaflet and the `MarkerService` from the application. It also defines the URLs for the marker icons and creates a default icon for the markers.

The component is decorated with `@Component` and has the following properties:
- `selector` set to `'app-map'`
- `templateUrl` set to `'./map.component.html'`
- `styleUrls` set to `['./map.component.sass']`
- `standalone` set to `true`

The `MapComponent` class implements the `AfterViewInit` interface and has a private `map` property.

The `initMap` method initializes the map using Leaflet