🎨 Add map component HTML file (master)
Added a new file `map.component.html` to the `src/app/map` directory. The HTML file contains a map container with a map frame and a map element. This will allow for the display of a map in the application.

This change enhances the application by providing the necessary HTML structure for the map component, enabling the integration of maps into the user interface.