🎨 (src/app/home/home.component.ts): Improve code formatting and import MapComponent
The code author made changes to the HomeComponent in order to improve its formatting and import the MapComponent. The code formatting was improved by aligning the import statements and adding line breaks. The MapComponent was imported to be used in the template. The template was also updated to include the MapComponent element. These changes aim to enhance the readability and functionality of the HomeComponent.