✨ :sparkles: Add makeSchoolPopup method to PopupService (master)
Add a new method called `makeSchoolPopup` to the `PopupService` class in the `popup.service.ts` file. This method returns a string that represents a popup with information about a school. The popup includes the school's name, address, type, and public/private status. This method will be used to generate popups for schools on the map.

This change enhances the functionality of the `PopupService` by allowing it to generate popups specifically for schools.