:sparkles: Add makeCapitalMarkers method to MarkerService
This commit adds a new method `makeCapitalMarkers` to the MarkerService class. This method is used to create markers for schools on a map. It retrieves school data from an API and creates a marker for each school on the map.

The `makeCapitalMarkers` method takes a map as a parameter and makes an HTTP request to retrieve school data. It then iterates over the data and creates a marker for each school. Each marker is bound to a popup using the `popupService.makeSchoolPopup` method. Finally, the marker is added to the map.

By adding this method, the code author aims to enhance the functionality of the MarkerService by allowing the display of school markers on a map. This will provide users with visual information about the location of schools in a specific area.