✨ Add Leaflet map component (master)
This commit adds an Angular component that displays a map using the Leaflet library. The component utilizes the MarkerService to add markers to the map. The map is initialized in the initMap() method by setting the center and zoom. Then, an OpenStreetMap tile layer is added to the map. The component implements the AfterViewInit interface to ensure that the map is initialized before adding the markers.

- Added Leaflet map component.
- Initialized the map with center and zoom.
- Added OpenStreetMap tile layer.
- Utilized MarkerService to add markers.
- Implemented AfterViewInit interface.