🔧 Refactor(map.component.sass): Adjust map container size.
The map container was previously set to occupy the entire page with a margin of 30px. However, this caused the map to overlap with other elements. With this change, the map container now occupies 100% of the width and 90% of the height, ensuring that it fits properly within the page layout.