:sparkles: Add Google Maps module and component to HomeComponent (master)
This commit adds the Google Maps module and component to the HomeComponent. The HomeComponent is responsible for displaying the home page of the application. By importing the GoogleMaps module and using the GoogleMaps component, a map is displayed on the home page.

The changes were made to the src/app/home/home.component.ts file.