Add modal and dialog services for terms and conditions component
A new TypeScript file 'tnc.component.ts' has been added to manage display of dialogs and modals in the application. Two services, DialogService and ModalService, are defined and utilized for opening dialogs and modals respectively. Included in the addition are definition of 'TnCContent' and 'TnCContentDialog' components. The former triggers a modal of terms and conditions on button click, while the latter holds the actual display content.